Refactor sign-up error messages
We used `user.errors.full_messages` which gives you an array of
human-readable error messages.

Plus we refactored the posts index view to get the posts's author's
email address directly and deleted the helper method. Neater, and we
delegated the `.email` method from the post through to the post's
author in accordance with the Law of Demeter.