Restrict automatically generated routes
We were automatically generating all routes for our users and posts
resources, some of which were unnecessary. So I've restricted the number
of routes generated to what we are actually using. This is in line with
Rails best practices, and cuts down on memory usage, speeding up the
routing process.

https://rails-bestpractices.com/posts/2011/08/19/restrict-auto-generated-routes/